requirements.txt
APScheduler==3.6.3
autopep8==1.5.5
backcall==0.2.0
CacheControl==0.12.6
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
decorator==4.4.2
firebase-admin==4.5.2
Flask==1.1.2
Flask-WTF==0.14.3
google-api-core==1.26.1
google-api-python-client==2.0.2
google-auth==1.27.1
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.0.2
google-cloud-storage==1.36.2
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpcio==1.36.1
gunicorn==20.0.4
httplib2==0.19.0
idna==2.10
ipdb==0.13.6
ipython==7.21.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
MarkupSafe==1.1.1
msgpack==1.0.2
packaging==20.9
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.17
proto-plus==1.17.0
protobuf==3.15.6
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.8.1
pyparsing==2.4.7
python-decouple==3.4
python-telegram-bot==13.4.1
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.4
wcwidth==0.2.5
Werkzeug==1.0.1
WTForms==2.3.3
